rearrange alarms
