+++ README.md
@@ -22,6 +22,8 @@ This repo uses the [Main-QA-Production](https://github.com/NYPL/engineering-gene
 
 `main` has the latest and greatest commits, `qa` has what's in our QA environment, and `production` has what's in our production environment.
 
+## Local Development
+
 ## Deployment
 CI/CD is not enabled. To deploy a new version of this function, first modify the code in the git repo and open a pull request to the appropriate environment branch. Then run `source deployment_script.sh` and upload the resulting zip. Note that if any files are added or deleted, this script must be modified. For more information, see the directions [here](https://docs.aws.amazon.com/lambda/latest/dg/python-package.html).
